Some clean up to the classinfo generator.
