+++ hexa.binding.gwt/src/main/java/fr/lteconsulting/hexa/classinfo/gwt/ClassInfoGwt.java
@@ -35,8 +35,6 @@ public class ClassInfoGwt implements IClassInfo
 
 	/**
 	 * Register a runtime type information provider
-	 * 
-	 * @param clazz
 	 */
 	public <T> void RegisterClazz( Clazz<T> clazz )
 	{
